+++ src/components/users/Users.jsx
@@ -6,7 +6,7 @@ import MaterialIcons from "react-native-vector-icons/MaterialIcons";
 
 const Users = ({ data, color, enable, dataConfiguration }) => {
   const navigation = useNavigation();
-  console.log(data); //! cuando se exporta despues se entra a detalle y vuelves a la vista de los usuarios existe un error (sucede cuando se exporta)
+
   // estilos dinámico del ícono de alerta
   const [estilos, setEstilos] = useState({
     fontSize: 40,
